+++ src/main/java/com/example/ecommercebe/entity/CartItem.java
@@ -27,9 +27,9 @@ public class CartItem {
 
     private int quantity;
 
-    private Integer price;
+    private Long price;
 
-    private Integer discountedPrice;
+    private Long discountedPrice;
 
     private Long userId;
